created components to cut down code size
