+++ client/src/pages/Login.jsx
@@ -1,128 +1,17 @@
-import { useState, useEffect } from "react";
-import { LockClosedIcon } from "@heroicons/react/solid";
-import { toast } from "react-toastify";
-import { login, reset } from "../features/auth/authSlice";
-import { useNavigate } from "react-router-dom";
-import { useSelector, useDispatch } from "react-redux";
-import Logo from "../assets/images/Logo.svg";
+// Components
+import LogoText from "../components/LogoText";
+import LoginForm from "../components/LoginForm";
 
 const Login = () => {
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-  });
-
-  const { email, password } = formData;
-
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-
-  const { user, isLoading, isSuccess, isError, message } = useSelector(
-    (state) => state.auth
-  );
-
-  useEffect(() => {
-    if (isError) {
-      toast.error(message);
-    }
-
-    if (isSuccess || user) {
-      navigate("/");
-    }
-
-    dispatch(reset());
-  }, [user, isError, isSuccess, message, navigate, dispatch]);
-
-  const onChange = (event) => {
-    setFormData((prevState) => ({
-      ...prevState,
-      [event.target.name]: event.target.value,
-    }));
-  };
-
-  const onSubmit = (event) => {
-    event.preventDefault();
-
-    const userData = {
-      email,
-      password,
-    };
-
-    dispatch(login(userData));
-  };
-
-  // Spinner
   return (
     <>
-      <div
-        onSubmit={onSubmit}
-        className="min-h-full flex items-center justify-center mt-48 py-12 px-4 sm:px-6 lg:px-8"
-      >
-        <div className="max-w-md w-full space-y-8">
-          <div>
-            <div className="flex justify-center items-center mb-5">
-              <img className="h-16 w-auto" src={Logo} alt="Track" />
-              <h2 className="font-bold h-15 text-5xl ml-4 italic text-gray-900 ">
-                Tracked{" "}
-              </h2>
-            </div>
-
-            <h2 className="mt-1 text-center text-3xl font-extrabold text-gray-900">
-              Sign In
-            </h2>
-          </div>
-          <form className="mt-8 space-y-6" onSubmit={onSubmit}>
-            <input type="hidden" name="remember" defaultValue="true" />
-            <div className="rounded-md shadow-sm -space-y-px">
-              <div>
-                <label htmlFor="email-address" className="sr-only">
-                  Email address
-                </label>
-                <input
-                  id="email-address"
-                  name="email"
-                  type="email"
-                  value={email}
-                  autoComplete="email"
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Email address"
-                  onChange={onChange}
-                />
-              </div>
-
-              <div>
-                <label htmlFor="password" className="sr-only">
-                  Password
-                </label>
-                <input
-                  id="password"
-                  name="password"
-                  type="password"
-                  value={password}
-                  autoComplete="current-password"
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Password"
-                  onChange={onChange}
-                />
-              </div>
-            </div>
-            <div>
-              <button
-                type="submit"
-                className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-teal-600 hover:bg-teal-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-teal-500"
-              >
-                <span className="absolute left-0 inset-y-0 flex items-center pl-3">
-                  <LockClosedIcon
-                    className="h-5 w-5 text-teal-500 group-hover:text-teal-400"
-                    aria-hidden="true"
-                  />
-                </span>
-                Sign in
-              </button>
-            </div>
-          </form>
+      <div className="min-h-full flex justify-center items-center mt-48 py-12 px-4 ">
+        <div className="max-w-md w-full space-y-4">
+          <LogoText />
+          <h2 className="text-center text-3xl font-extrabold text-gray-900 mt-1">
+            Sign In
+          </h2>
+          <LoginForm />
         </div>
       </div>
     </>
