+++ client/src/pages/Register.jsx
@@ -1,169 +1,17 @@
-import { useState, useEffect } from "react";
-import { LoginIcon } from "@heroicons/react/outline";
-import Logo from "../assets/images/Logo.svg";
-import { Link } from "react-router-dom";
-import { toast } from "react-toastify";
-import { register, reset } from "../features/auth/authSlice";
-import { useNavigate } from "react-router-dom";
-import { useSelector, useDispatch } from "react-redux";
+// Components
+import LogoText from "../components/LogoText";
+import RegisterForm from "../components/RegisterForm";
 
 const Register = () => {
-  const [formData, setFromData] = useState({
-    name: "",
-    email: "",
-    password: "",
-    passwordConfirm: "",
-  });
-
-  const { name, email, password, passwordConfirm } = formData;
-
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-
-  const { user, isLoading, isSuccess, isError, message } = useSelector(
-    (state) => state.auth
-  );
-
-  useEffect(() => {
-    if (isError) {
-      toast.error(message);
-    }
-
-    if (isSuccess || user) {
-      navigate("/");
-    }
-
-    dispatch(reset());
-  }, [user, isError, isSuccess, message, navigate, dispatch]);
-
-  const onChange = (event) => {
-    setFromData((prevState) => ({
-      ...prevState,
-      [event.target.name]: event.target.value,
-    }));
-  };
-
-  const onSubmit = (event) => {
-    event.preventDefault();
-
-    if (password !== passwordConfirm) {
-      toast.error("Passwords do not match");
-    } else {
-      const userData = {
-        name,
-        email,
-        password,
-      };
-
-      dispatch(register(userData));
-    }
-  };
-
-  if (isLoading) {
-    // return spinner
-  }
-
   return (
     <>
-      <div className="min-h-full flex items-center justify-center mt-32 py-12 px-4 sm:px-6 lg:px-8">
+      <div className="min-h-full flex justify-center items-center mt-32 py-12 px-4 ">
         <div className="max-w-md w-full space-y-8">
-          <div>
-            <div className="flex justify-center items-center mb-5">
-              <img className="h-16 w-auto" src={Logo} alt="Track" />
-              <h2 className="font-bold h-15 text-5xl ml-4 italic text-gray-900 ">
-                Tracked{" "}
-              </h2>
-            </div>
-            <h2 className="mt-1 text-center text-3xl font-extrabold text-grey-900">
-              Register your account
-            </h2>
-          </div>
-          <form className="mt-8 space-y-6" onSubmit={onSubmit}>
-            <input type="hidden" name="remember" defaultValue="true" />
-            <div className="rounded-md shadow-sm -space-y-px">
-              <div>
-                <label htmlFor="email-address" className="sr-only">
-                  Name
-                </label>
-                <input
-                  id="name"
-                  name="name"
-                  type="text"
-                  value={name}
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Name"
-                  onChange={onChange}
-                />
-              </div>
-              <div>
-                <label htmlFor="email-address" className="sr-only">
-                  Email address
-                </label>
-                <input
-                  id="email-address"
-                  name="email"
-                  type="email"
-                  value={email}
-                  autoComplete="email"
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Email address"
-                  onChange={onChange}
-                />
-              </div>
-              <div>
-                <label htmlFor="password" className="sr-only">
-                  Password
-                </label>
-                <input
-                  id="password"
-                  name="password"
-                  type="password"
-                  value={password}
-                  autoComplete="current-password"
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Password"
-                  onChange={onChange}
-                />
-              </div>
-              <div>
-                <label htmlFor="password-confirm" className="sr-only">
-                  Confirm Password
-                </label>
-                <input
-                  id="password-confirm"
-                  name="passwordConfirm"
-                  type="password"
-                  value={passwordConfirm}
-                  required
-                  className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-teal-500 focus:border-teal-500 focus:z-10 sm:text-sm"
-                  placeholder="Confirm Password"
-                  onChange={onChange}
-                />
-              </div>
-            </div>
-            <div className="grid justify-items-end text-sm">
-              <span className="font-medium  text-teal-600 hover:text-teal-500">
-                <Link to="/login"> Already have an account?</Link>
-              </span>
-            </div>
-            <div>
-              <button
-                type="submit"
-                className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-teal-600 hover:bg-teal-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-teal-500"
-              >
-                <span className="absolute left-0 inset-y-0 flex items-center pl-3">
-                  <LoginIcon
-                    className="h-5 w-5 text-teal-500 group-hover:text-teal-400"
-                    aria-hidden="true"
-                  />
-                </span>
-                Sign in
-              </button>
-            </div>
-          </form>
+          <LogoText />
+          <h2 className="text-center text-3xl font-extrabold text-grey-900 mt-1 ">
+            Register your account
+          </h2>
+          <RegisterForm />
         </div>
       </div>
     </>
